v1.1リリース向けにmasterへマージ

+++ backendapp/services/JsonToDatabase.ts
@@ -211,6 +211,7 @@ export interface JSONSchema7 {
   'jesgo:valid'?: string[] | undefined;
   'jesgo:version'?: string | undefined;
   'jesgo:author'?: string | undefined;
+  'jesgo:inheriteventdate'?: string | undefined;
 }
 
 type oldSchema = {
@@ -730,10 +731,10 @@ const fileListInsert = async (
   fileList: string[],
   errorMessages: string[],
   dirPath: string
-): Promise<number> => {
+): Promise<{ updateNum: number; isSchemaAllUpdate: boolean }> => {
   logging(LOGTYPE.DEBUG, `呼び出し`, 'JsonToDatabase', 'fileListInsert');
   let updateNum = 0;
-  const jsons: JSONSchema7[] = [];
+  const jsons: { json: JSONSchema7; filePath: string }[] = [];
   for (let i = 0; i < fileList.length; i++) {
     if (!fileList[i].endsWith('.json')) {
       logging(
@@ -756,7 +757,7 @@ const fileListInsert = async (
     let json: JSONSchema7 = {};
     try {
       json = JSON.parse(readFileSync(fileList[i], 'utf8')) as JSONSchema7;
-      jsons.push(json);
+      jsons.push({ json, filePath: fileList[i] });
     } catch {
       logging(
         LOGTYPE.ERROR,
@@ -782,16 +783,16 @@ const fileListInsert = async (
     // 文字列の短い順(よりパスの短い順)に整列
     .sort(function (a, b) {
       // $ID順で並び替え
-      if (a.$id && b.$id) {
-        if (a.$id !== b.$id) {
-          if (a.$id > b.$id) return 1;
-          if (a.$id < b.$id) return -1;
+      if (a.json.$id && b.json.$id) {
+        if (a.json.$id !== b.json.$id) {
+          if (a.json.$id > b.json.$id) return 1;
+          if (a.json.$id < b.json.$id) return -1;
         }
       }
 
       // バージョンの比較
-      const versionOfA = a['jesgo:version'];
-      const versionOfB = b['jesgo:version'];
+      const versionOfA = a.json['jesgo:version'];
+      const versionOfB = b.json['jesgo:version'];
       if (versionOfA && versionOfB) {
         // メジャーバージョン順で並び替え
         if (
@@ -814,19 +815,31 @@ const fileListInsert = async (
       return 0;
     });
 
+  const insertedList: JSONSchema7[] = [];
+  let isSchemaAllUpdate = false;
+
   for (let i = 0; i < jsons.length; i++) {
     // Insert用IDを含む旧データの取得
-    const oldJsonData = await getOldSchema(jsons[i].$id as string);
+    const oldJsonData = await getOldSchema(jsons[i].json.$id as string);
 
     const [query, values, subqueryValues] = makeInsertQuery(
       oldJsonData,
-      jsons[i],
+      jsons[i].json,
       dirPath,
-      fileList[i],
+      jsons[i].filePath,
       errorMessages
     );
     if (query !== '') {
       await dbAccess.query(query, values);
+
+      // 同一スキーマIDのスキーマが処理された場合、後続のschemaListUpdateを全スキーマ更新モードで実行する
+      if (insertedList.find((p) => p.$id === jsons[i].json.$id)) {
+        isSchemaAllUpdate = true;
+      }
+      if (!isSchemaAllUpdate) {
+        insertedList.push(jsons[i].json);
+      }
+
       if (subqueryValues[0] !== '') {
         // 旧スキーマの有効期限更新がある場合そちらも行う
         await dbAccess.query(
@@ -837,13 +850,13 @@ const fileListInsert = async (
       updateNum++;
     }
   }
-  return updateNum;
+  return { updateNum, isSchemaAllUpdate };
 };
 
 /**
  * DBに登録されているスキーマのsubschema, childschema情報をアップデートする
  */
-export const schemaListUpdate = async () => {
+export const schemaListUpdate = async (updateAll = false) => {
   logging(LOGTYPE.DEBUG, `呼び出し`, 'JsonToDatabase', 'schemaListUpdate');
 
   // 先に「子スキーマから指定した親スキーマの関係リスト」を逆にしたものを作成しておく
@@ -917,6 +930,7 @@ export const schemaListUpdate = async () => {
 
   type dbRow = {
     schema_id: number;
+    schema_primary_id: number;
     schema_id_string: string;
     sub_s: string[];
     child_s: string[];
@@ -928,14 +942,17 @@ export const schemaListUpdate = async () => {
   // selectでDBに保存されている各スキーマのschema_id,schema_string_id,subschema,childschema一覧を取得
   const dbRows: dbRow[] = (await dbAccess.query(
     `SELECT schema_id, 
+    schema_primary_id,
     schema_id_string, 
     document_schema->'jesgo:subschema' as sub_s, 
     document_schema->'jesgo:childschema' as child_s, 
     subschema_default as default_sub_s, 
     child_schema_default as default_child_s 
-    FROM view_latest_schema 
+    FROM ${updateAll ? 'jesgo_document_schema' : 'view_latest_schema'} 
     WHERE schema_id <> 0`
   )) as dbRow[];
+
+  let updateCount = 0;
 
   const candidateBaseSchemas = dbRows.slice(0);
   for (let i = 0; i < dbRows.length; i++) {
@@ -944,6 +961,7 @@ export const schemaListUpdate = async () => {
     const childSchemaList: number[] = [];
     const inheritSchemaList: number[] = [];
     let baseSchemaId: number | undefined;
+    // subschema
     if (row.sub_s != null) {
       for (let j = 0; j < row.sub_s.length; j++) {
         // ワイルドカードを含むかどうかで処理を分ける
@@ -958,7 +976,7 @@ export const schemaListUpdate = async () => {
             [searchId]
           )) as schemaId[];
           if (schemaIds.length > 0) {
-            subSchemaList.push(schemaIds[0].schema_id);
+            subSchemaList.push(...schemaIds.map((p) => p.schema_id));
           }
         } else {
           const schemaIds: schemaId[] = (await dbAccess.query(
@@ -971,6 +989,7 @@ export const schemaListUpdate = async () => {
         }
       }
     }
+    // child_schema
     if (row.child_s != null) {
       for (let k = 0; k < row.child_s.length; k++) {
         // ワイルドカードを含むかどうかで処理を分ける
@@ -985,7 +1004,7 @@ export const schemaListUpdate = async () => {
             [searchId]
           )) as schemaId[];
           if (schemaIds.length > 0) {
-            childSchemaList.push(schemaIds[0].schema_id);
+            childSchemaList.push(...schemaIds.map((p) => p.schema_id));
           }
         } else {
           const schemaIds: schemaId[] = (await dbAccess.query(
@@ -1065,9 +1084,13 @@ export const schemaListUpdate = async () => {
          , child_schema_default = '{${numArrayCast2Pg(newChildSchemaList)}}'`;
     }
 
-    query += ' WHERE schema_id = $1';
+    query += ' WHERE schema_primary_id = $1';
 
-    await dbAccess.query(query, [row.schema_id]);
+    updateCount += (await dbAccess.query(
+      query,
+      [row.schema_primary_id],
+      'update'
+    )) as number;
   }
 
   // スキーマの更新に合わせて検索用セレクトボックスも更新する
@@ -1075,6 +1098,8 @@ export const schemaListUpdate = async () => {
 
   // スキーマの更新に合わせてルートスキーマの内容も更新する
   await updateRootSchemaList();
+
+  return updateCount;
 };
 
 const updateRootSchemaList = async () => {
@@ -1154,10 +1179,10 @@ export const updateSearchColumn = async (): Promise<void> => {
     `SELECT document_schema 
     FROM view_latest_schema 
     WHERE document_schema->>'properties' like '%${escapeText(
-      jesgo_tagging(Const.JESGO_TAG.CANCER_MAJOR)
+      `"${Const.JESGO_TAG.CANCER_MAJOR}"`
     )}%' 
     OR document_schema->>'properties' like '%${escapeText(
-      jesgo_tagging(Const.JESGO_TAG.CANCER_MINOR)
+      `"${Const.JESGO_TAG.CANCER_MINOR}"`
     )}%' 
     AND schema_id <> 0 
     ORDER BY schema_id_string;`
@@ -1192,17 +1217,28 @@ export const updateSearchColumn = async (): Promise<void> => {
     }
   }
 
-  // 取得した情報でDBを更新
-  await dbAccess.query(
-    "DELETE FROM jesgo_search_column WHERE column_type = 'cancer_type'"
-  );
   const cancerList = majorCancers.concat(minorCancers);
 
-  for (let i = 0; i < cancerList.length; i++) {
+  // 現在のリストを取得
+  const searchColumnResult = (await dbAccess.query(
+    `SELECT column_name FROM jesgo_search_column ORDER BY column_id`
+  )) as { column_name: string }[];
+  const searchColumnList = searchColumnResult.map((p) => p.column_name);
+
+  // 現在のリストとスキーマから生成したリストが異なれば更新する
+  if (JSON.stringify(searchColumnList) !== JSON.stringify(cancerList)) {
     await dbAccess.query(
-      "INSERT INTO jesgo_search_column VALUES ($1, 'cancer_type', $2)",
-      [i + 1, cancerList[i]]
+      "DELETE FROM jesgo_search_column WHERE column_type = 'cancer_type'",
+      undefined,
+      'update'
     );
+
+    for (let i = 0; i < cancerList.length; i++) {
+      await dbAccess.query(
+        "INSERT INTO jesgo_search_column VALUES ($1, 'cancer_type', $2)",
+        [i + 1, cancerList[i]]
+      );
+    }
   }
 };
 
@@ -1300,16 +1336,16 @@ export const uploadZipFile = async (data: any): Promise<ApiReturnObject> => {
 
     await dbAccess.connectWithConf();
 
-    const updateNum = await fileListInsert(fileList, errorMessages, dirPath);
+    const result = await fileListInsert(fileList, errorMessages, dirPath);
 
     // スキーマが1件以上新規登録、更新された場合のみ関係性のアップデートを行う
-    if (updateNum > 0) {
-      await schemaListUpdate();
+    if (result.updateNum > 0) {
+      await schemaListUpdate(result.isSchemaAllUpdate);
     }
 
     return {
       statusNum: RESULT.NORMAL_TERMINATION,
-      body: { number: updateNum, message: errorMessages },
+      body: { number: result.updateNum, message: errorMessages },
     };
   } catch (e) {
     if (dbAccess.connected) {
@@ -1376,5 +1412,43 @@ export const uploadZipFile = async (data: any): Promise<ApiReturnObject> => {
         'uploadZipFile'
       );
     });
+  }
+};
+
+/**
+ * subschema、child_schemaのみ更新
+ * @returns
+ */
+export const repairChildSchema = async (): Promise<ApiReturnObject> => {
+  logging(LOGTYPE.DEBUG, `呼び出し`, 'JsonToDatabase', 'repairChildSchema');
+
+  try {
+    await dbAccess.connectWithConf();
+    await dbAccess.query('BEGIN');
+
+    const count = await schemaListUpdate(true);
+
+    await dbAccess.query('COMMIT');
+    return {
+      statusNum: RESULT.NORMAL_TERMINATION,
+      body: `subschema、child_schemaの更新に成功しました。(${
+        count ?? ''
+      }件の更新)`,
+    };
+  } catch (e) {
+    logging(
+      LOGTYPE.ERROR,
+      `${(e as Error).message}`,
+      'JsonToDatabase',
+      'repairChildSchema'
+    );
+    await dbAccess.query('ROLLBACK');
+    return {
+      statusNum: RESULT.ABNORMAL_TERMINATION,
+      body: 'subschema、child_schemaの更新に失敗しました',
+      error: (e as Error).message,
+    };
+  } finally {
+    await dbAccess.end();
   }
 };
